pyproject.toml
[tool.poetry]
name = "algobot"
version = "0.1.0"
description = "Web application for analyzing and tracking cryptocurrency prices using AI and machine learning techniques"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.3.3"
Flask-Cors = "^4.0.1"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.7"
Flask-RESTx = "^1.3.0"
Flask-JWT-Extended = "^4.6.0"
Flask-Mail = "^0.9.1"
Flask-Caching = "^2.1.0"
SQLAlchemy = "^2.0.32"
alembic = "^1.13.2"
gunicorn = "^21.2.0"
celery = "^5.4.0"
redis = "^5.0.8"
marshmallow = "^3.21.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
numpy = "^1.26.4"
scikit-learn = "^1.5.1"
matplotlib = "^3.9.2"
plotly = "^5.23.0"
dash = "^2.17.1"
beautifulsoup4 = "^4.12.3"
tweepy = "^4.14.0"
praw = "^7.7.1"
alpaca-trade-api = "^3.2.0"
fredapi = "^0.5.2"
openai = "^1.41.0"
APScheduler = "^3.10.4"
textblob = "^0.17.1"
geckoterminal = "^0.0.6"
ccxt = "^4.3.84"
ta = "^0.11.0"
flower = "^2.0.1"
marshmallow-sqlalchemy = "^1.1.0"
nltk = "^3.9.1"
cachetools = "^5.5.0"
diskcache = "^5.6.3"
pytest-asyncio = "^0.23.8"
asgiref = "^3.8.1"
hypercorn = "^0.17.3"
polygon-api-client = "^1.14.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
flake8 = "^7.1.1"
black = "^23.12.1"
mypy = "^1.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "flask run"
lint = "flake8 app tests"
format = "black app tests && isort app tests"
test = "pytest tests"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]